requirement.txt
django
djangorestframework
djangorestframework-jsonapi
django-filter
django-cors-headers
pillow